Simplify evaluation semantics
Since in python the only thing you can do is raise an exception
(sys.exit also just raises an exception), offering auto-exiting
semantics is kind of pointless.

Also, I've since come around on whether an exception escaping from a
predicate is a contract failure or not. The new answer is no - that
exception just escapes and is handled according to the usual language
rules.